Revert to S3 method system
This change comes after having gone first from S3 to S4 then from S4 to R6. There is obviously a swath of changes to the codebase. The motivation boils down to making documentation easier, which has been a pain when working with R6 and also make debugging easier (for the very same reason).